[processing] Move AlgorithmTests to QgisAlgorithmTests
