+++ python/plugins/processing/tests/CMakeLists.txt
@@ -7,5 +7,5 @@ PLUGIN_INSTALL(processing tests/data ${TEST_DATA_FILES})
 IF(ENABLE_TESTS)
   INCLUDE(UsePythonTest)
   ADD_PYTHON_TEST(ProcessingParametersTest ParametersTest.py)
-  ADD_PYTHON_TEST(ProcessingAlgorithmsTest AlgorithmsTest.py)
+  ADD_PYTHON_TEST(ProcessingQgisAlgorithmsTest QgisAlgorithmsTest.py)
 ENDIF(ENABLE_TESTS)
